refactor(store): Store와 StoreName 객체의 패키지 이동 (#91)
- Store와 StoreName 객체의 패키지 이동
- 다른 클래스들의 importing 반영
