+++ src/test/java/com/firstone/greenjangteo/user/domain/store/service/model/StoreNameTest.java
@@ -1,5 +1,6 @@
 package com.firstone.greenjangteo.user.domain.store.service.model;
 
+import com.firstone.greenjangteo.user.domain.store.model.StoreName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
