+++ src/test/java/com/firstone/greenjangteo/user/domain/store/testutil/TestObjectFactory.java
@@ -1,7 +1,7 @@
 package com.firstone.greenjangteo.user.domain.store.testutil;
 
 import com.firstone.greenjangteo.user.domain.store.dto.StoreDto;
-import com.firstone.greenjangteo.user.domain.store.service.model.entity.Store;
+import com.firstone.greenjangteo.user.domain.store.model.entity.Store;
 
 public class TestObjectFactory {
     public static Store createStore(Long sellerId, String storeName, String description, String imageUrl) {
